Add search filters
Handle filters passed to the Search LDAP method. The filter is parsed
and validated and run against all the entries selected by the scope and
the base DN. All entries that match the filter are returned.

The full set of LDAP filter are not implemented. Currently implemented
are:
* Presence filters
* Equality filters
* And / Or / Not filters

Filters not implemented yet are:
* Greater-Or-Equal / Less-Or-Equal filters
* Substring filters
* Approximate Match filters
* Extensible Match filters