✨ Add basic server and simple database loaded from jsonnet (#3)
Add a basic server that can serve the 'Bind' and `Search` LDAP functions.

The Bind method simply logs its parameters in order to see exactly how
they arrive, testing using `ldapsearch`. The logging will be replace
with actual authentication when implemented.

The Search method is backed by a simple database. The database is loaded
from a jsonnet file (which can be just json if desired). Add anonymous
binds, and a stub for password binds.

The database is currently case-sensitive, and is not backed by a schema.
It is a tree made out of `DITNode` (Directory Information Tree nodes),
each having an `Entry` and child nodes. Each `Entry` has a Distinguished
Name (DN) and a map of attribute names to slice of values. Values are of
type string only.

This is a far cry from a proper LDAP database, but serves well as a
stand-in during development. Future changes may add case-insensitive
attribute names, and once a schema is added, the attribute values can be
typed and possibly case-insensitive too.

Disable a couple of gocritic linters because they're stupid.

Pull-request: https://github.com/foxygoat/flapjak/pull/3