✨ Add simple bind authentication (#9)
Handle simple bind requests of the server using the username and
password fields. The username is the DN of an entry that must be of an
appropriate `objectClass` for authentication and contain a
`userPassword` attribute containing a value with a supported hash
scheme. The password passed to the bind method is hashed according to
the scheme and compared to the stored hash, and if it matches, the bind
method returns success.

Pull-request: https://github.com/foxygoat/flapjak/pull/9