Only return Directory Server Entry object if explicitly requested
Ensure we return the Directory Server Entry (DSE) when the requested DN
is the empty DN and the scope is for a base object (not a tree).

The Directory Server Entry (DSE) is the root of the DB and contains
server/service oriented information, not user data. This should only be
returned when explicitly requested by requesting the empty DN and
requesting just a base object (not a tree).

Previously we could get the DSE based on a search for a DN that is a
parent of the root data entry and we'd return that.