Add container image build
Add the `build-oci` target to the `Makefile` to build a container image
and push it to a registry. Call the target from the `release` target so
that container images are now part of the release.

The registry, target architectures (platforms) and image tag can be
customised on a per-environment basis by defining variables for these
values with the environment name suffix. An environment for "release" is
defined in the makefile to be used by the `release` target to publish
images to docker hub. Local configuration (perhaps via direnv) allows
other environments to be defined so that, for example, images can be
pushed to a local registry for development and testing.

Add a `docker-login` target to the `Makefile` so the release automation
can set up the appropriate credentials for pushing to a registry.