🐛 ci: Enable containerd snapshotter on GitHub Actions (#12)
Enable the containerd snapshotter on the Docker daemon on the GitHub
Actions runner so we have a multi-architecture image store and so that
`docker buildx` will work. We do cross compiling with Go so we do not
need the qemu thing for running the builder on a foreign architecture.

Test-run: https://github.com/foxygoat/flapjak/actions/runs/19528620262
Pull-request: https://github.com/foxygoat/flapjak/pull/12