tests: Convert stretchr/testify to matryer/is
Convert all the tests from using `stretchr/testify` to `matryer/is`. `is`
is a much smaller package with fewer dependencies. Most of the
additional functionality of testify can be replaced with direct Go code
or the `google/go-cmp` package (which we do not need here yet).

Fix the comment on `IsAncestor()` which had some bits backwards. Also
fix the test comment for `IsAncestor()` which had the same backwards.